requirements.txt
asgiref==3.5.0
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3
Django==4.0.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.4.1
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
